requirements.txt
werkzeug
progress
qrcode
pyzbar
Pillow